examples/agno/requirements.txt
# Requirements for Agno Tool Integration Examples
# Install with: pip install -r requirements.txt

# Core Agno framework
agno>=1.7.0

# AI/ML APIs
openai>=1.93.0
cohere>=5.15.0

# Vector database and embeddings
lancedb>=0.24.0
numpy>=2.3.0
pyarrow>=16.0.0

# Monitoring and observability
agentops>=0.4.17

# Environment and configuration
python-dotenv>=1.1.0

# HTTP and networking
httpx>=0.28.0
requests>=2.32.0

# Data processing
pyyaml>=6.0.0
pydantic>=2.11.0

# Optional: Additional dependencies for enhanced functionality
# uncomment if needed:
# chromadb>=0.4.0  # Alternative vector database
# pinecone-client>=3.0.0  # Alternative vector database
# sentence-transformers>=2.2.0  # Alternative embeddings